pyproject.toml
[tool.poetry]
name = "ecgai-drawing-azure-httpfunction"
version = "0.1.0"
description = ""
authors = ["RobC <rob.clapham@gmail.com>"]

[tool.poetry.dependencies]
#TODO change poetry version to <4.0, currently required by scipy
python = "^3.9,<3.12"
fastapi = "^0.85.0"
uvicorn = "^0.18.3"
azure-functions = "^1.11.3"
ecgai-drawing = {git = "https://github.com/Ecg-Ai-com/ecgai-drawing.git", rev = "main"}
nest-asyncio = "^1.5.6"

[tool.poetry.dev-dependencies]
pre-commit = "^2.20.0"
pydevd-pycharm = "^223.6160.21"
pytest = "^7.1.3"
pytest-asyncio = "^0.19.0"
#azure-functions-worker = "1.0.0a6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.black]
line-length = 120
